requirements.txt
python-telegram-bot>=21.0
psutil>=5.9.0
pyautogui>=0.9.54
pyperclip>=1.8.2
mss>=9.0.1
requests>=2.31.0
python-dotenv>=1.0.0
APScheduler>=3.10.4
playwright>=1.44.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
